Radius type & replace DiscountsDAOImpl source to ShopsDAOImpl

+++ src/alexiuscrow/diploma/dao/ShopsDAO.java
@@ -7,6 +7,6 @@ import alexiuscrow.diploma.entity.Shops;
 
 public interface ShopsDAO {    
     public List<Shops> getAllShops() throws SQLException;
-	public String getNearestShops(Double lat, Double lng, Integer radius) throws SQLException;
+	public String getNearestShops(Double lat, Double lng, Double radius) throws SQLException;
 	public String getAllLocalityShops(Double lat, Double lng) throws SQLException;
 }
